# api-docs.json
{"openapi":"3.1.0","info":{"title":"The Great EIA Camel vs. Dwarf Racing API","description":"API REST para la liga de carreras Camel vs. Dwarf.\n\nAutenticación con Keycloak: usa el botón **Authorize** e\ninicia sesión con uno de los usuarios de prueba del realm.\n","contact":{"name":"Nheo Automatization Team"},"license":{"name":"Uso académico"},"version":"0.0.1-SNAPSHOT"},"servers":[{"url":"http://localhost:8080","description":"Entorno local"}],"security":[{"keycloak":[]},{"bearer-token":[]}],"paths":{"/api/teams/{id}":{"get":{"tags":["team-controller"],"operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}}}},"put":{"tags":["team-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}}}},"delete":{"tags":["team-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/results/{id}":{"get":{"tags":["race-result-controller"],"operationId":"findById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaceResultResponse"}}}}}},"put":{"tags":["race-result-controller"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RaceResultRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaceResultResponse"}}}}}}},"/api/races/{id}":{"get":{"tags":["race-controller"],"operationId":"findById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaceResponse"}}}}}},"put":{"tags":["race-controller"],"operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RaceRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaceResponse"}}}}}},"delete":{"tags":["race-controller"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/competitors/{id}":{"get":{"tags":["competitor-controller"],"operationId":"getCompetitorById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CompetitorResponse"}}}}}},"put":{"tags":["competitor-controller"],"operationId":"updateCompetitor","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompetitorRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CompetitorResponse"}}}}}},"delete":{"tags":["competitor-controller"],"operationId":"deleteCompetitor","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/teams":{"get":{"tags":["team-controller"],"operationId":"findAll","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["ACTIVE","SUSPENDED","INACTIVE"]}},{"name":"coach","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"name"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"asc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageTeamResponse"}}}}}},"post":{"tags":["team-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}}}}},"/api/teams/{teamId}/members/{competitorId}":{"post":{"tags":["team-controller"],"operationId":"addMember","parameters":[{"name":"teamId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"competitorId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}}}},"delete":{"tags":["team-controller"],"operationId":"removeMember","parameters":[{"name":"teamId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"competitorId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/races":{"get":{"tags":["race-controller"],"operationId":"findAll_1","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"raceType","in":"query","required":false,"schema":{"type":"string","enum":["INDIVIDUAL","TEAM","MIXED"]}},{"name":"raceStatus","in":"query","required":false,"schema":{"type":"string","enum":["DRAFT","OPEN_FOR_REGISTRATION","CLOSED_FOR_REGISTRATION","IN_PROGRESS","COMPLETED","CANCELLED"]}},{"name":"fromDate","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"toDate","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"organizer","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"scheduledDateTime"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"asc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageRaceResponse"}}}}}},"post":{"tags":["race-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RaceRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaceResponse"}}}}}}},"/api/races/{raceId}/results":{"get":{"tags":["race-result-controller"],"operationId":"findAllByRace","parameters":[{"name":"raceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RaceResultResponse"}}}}}}},"post":{"tags":["race-result-controller"],"operationId":"create_2","parameters":[{"name":"raceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RaceResultRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaceResultResponse"}}}}}}},"/api/races/{raceId}/registrations":{"get":{"tags":["race-registration-controller"],"operationId":"findAllByRace_1","parameters":[{"name":"raceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RaceRegistrationResponse"}}}}}}},"post":{"tags":["race-registration-controller"],"operationId":"create_3","parameters":[{"name":"raceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RaceRegistrationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaceRegistrationResponse"}}}}}}},"/api/competitors":{"get":{"tags":["competitor-controller"],"operationId":"getAllCompetitors","parameters":[{"name":"type","in":"query","required":false,"schema":{"type":"string","enum":["DWARF","CAMEL","MEDIUM","OTHER"]}},{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["ACTIVE","INJURED","SUSPENDED","RETIRED"]}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCompetitorResponse"}}}}}},"post":{"tags":["competitor-controller"],"operationId":"createCompetitor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompetitorRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CompetitorResponse"}}}}}}},"/api/auth/refresh":{"post":{"tags":["auth-controller"],"summary":"Cambia un refresh token vigente por un access token nuevo","operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"summary":"Inicia sesión contra Keycloak y devuelve el access/refresh token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/teams/{id}/status":{"patch":{"tags":["team-controller"],"operationId":"updateStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTeamStatusRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}}}}},"/api/registrations/{id}/reject":{"patch":{"tags":["race-registration-controller"],"operationId":"reject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RejectRegistrationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaceRegistrationResponse"}}}}}}},"/api/registrations/{id}/approve":{"patch":{"tags":["race-registration-controller"],"operationId":"approve","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaceRegistrationResponse"}}}}}}},"/api/races/{id}/status":{"patch":{"tags":["race-controller"],"operationId":"updateStatus_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRaceStatusRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaceResponse"}}}}}}},"/api/competitors/{id}/status":{"patch":{"tags":["competitor-controller"],"operationId":"updateCompetitorStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStatusRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CompetitorResponse"}}}}}}},"/api/standings":{"get":{"tags":["standings-controller"],"operationId":"getStandings","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandingsResponse"}}}}}}},"/api/standings/teams":{"get":{"tags":["standings-controller"],"operationId":"getTeamStandings","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TeamStandingResponse"}}}}}}}},"/api/standings/competitors":{"get":{"tags":["standings-controller"],"operationId":"getCompetitorStandings","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CompetitorStandingResponse"}}}}}}}},"/api/registrations/{id}":{"get":{"tags":["race-registration-controller"],"operationId":"findById_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaceRegistrationResponse"}}}}}},"delete":{"tags":["race-registration-controller"],"operationId":"cancel","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/auth/profile":{"get":{"tags":["auth-controller"],"summary":"Devuelve la identidad y los roles del usuario autenticado","operationId":"profile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}}}}}},"components":{"schemas":{"TeamRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0},"description":{"type":"string","maxLength":500,"minLength":0},"coach":{"type":"string","maxLength":100,"minLength":0}},"required":["coach","name"]},"MemberResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"nickname":{"type":"string"}}},"TeamResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"creationDate":{"type":"string","format":"date"},"coach":{"type":"string"},"status":{"type":"string","enum":["ACTIVE","SUSPENDED","INACTIVE"]},"victories":{"type":"integer","format":"int32"},"defeats":{"type":"integer","format":"int32"},"memberCount":{"type":"integer","format":"int32"},"members":{"type":"array","items":{"$ref":"#/components/schemas/MemberResponse"}}}},"RaceResultRequest":{"type":"object","properties":{"registrationId":{"type":"integer","format":"int64"},"resultStatus":{"type":"string","enum":["FINISHED","DISQUALIFIED","DID_NOT_FINISH","DID_NOT_START"]},"finalPosition":{"type":"integer","format":"int32"},"completionTimeMillis":{"type":"integer","format":"int64"},"penaltyTimeMillis":{"type":"integer","format":"int64"},"notes":{"type":"string","maxLength":500,"minLength":0},"recordedBy":{"type":"string","maxLength":120,"minLength":0}},"required":["penaltyTimeMillis","recordedBy","registrationId","resultStatus"]},"RaceResultResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"raceId":{"type":"integer","format":"int64"},"raceName":{"type":"string"},"registrationId":{"type":"integer","format":"int64"},"participantType":{"type":"string"},"competitorId":{"type":"string","format":"uuid"},"competitorName":{"type":"string"},"teamId":{"type":"integer","format":"int64"},"teamName":{"type":"string"},"startingPosition":{"type":"integer","format":"int32"},"finalPosition":{"type":"integer","format":"int32"},"completionTimeMillis":{"type":"integer","format":"int64"},"penaltyTimeMillis":{"type":"integer","format":"int64"},"totalTimeMillis":{"type":"integer","format":"int64"},"resultStatus":{"type":"string","enum":["FINISHED","DISQUALIFIED","DID_NOT_FINISH","DID_NOT_START"]},"points":{"type":"integer","format":"int32"},"notes":{"type":"string"},"recordedBy":{"type":"string"},"recordedAt":{"type":"string","format":"date-time"}}},"RaceRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":150,"minLength":0},"description":{"type":"string","maxLength":500,"minLength":0},"scheduledDateTime":{"type":"string","format":"date-time"},"startLocation":{"type":"string","maxLength":150,"minLength":0},"finishLocation":{"type":"string","maxLength":150,"minLength":0},"distanceMeters":{"type":"number","format":"double"},"maxParticipants":{"type":"integer","format":"int32"},"raceType":{"type":"string","enum":["INDIVIDUAL","TEAM","MIXED"]},"organizer":{"type":"string","maxLength":120,"minLength":0},"registrationDeadline":{"type":"string","format":"date-time"}},"required":["finishLocation","name","organizer","raceType","registrationDeadline","scheduledDateTime","startLocation"]},"RaceResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"scheduledDateTime":{"type":"string","format":"date-time"},"startLocation":{"type":"string"},"finishLocation":{"type":"string"},"distanceMeters":{"type":"number","format":"double"},"maxParticipants":{"type":"integer","format":"int32"},"raceType":{"type":"string","enum":["INDIVIDUAL","TEAM","MIXED"]},"raceStatus":{"type":"string","enum":["DRAFT","OPEN_FOR_REGISTRATION","CLOSED_FOR_REGISTRATION","IN_PROGRESS","COMPLETED","CANCELLED"]},"organizer":{"type":"string"},"registrationDeadline":{"type":"string","format":"date-time"},"creationDate":{"type":"string","format":"date-time"},"lastModificationDate":{"type":"string","format":"date-time"}}},"CompetitorRequest":{"type":"object","properties":{"name":{"type":"string"},"nickname":{"type":"string"},"competitorType":{"type":"string","enum":["DWARF","CAMEL","MEDIUM","OTHER"]},"age":{"type":"integer","format":"int32"},"weight":{"type":"number","format":"double"},"height":{"type":"number","format":"double"},"country":{"type":"string"}},"required":["competitorType","country","name","nickname"]},"CompetitorResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"nickname":{"type":"string"},"competitorType":{"type":"string","enum":["DWARF","CAMEL","MEDIUM","OTHER"]},"age":{"type":"integer","format":"int32"},"weight":{"type":"number","format":"double"},"height":{"type":"number","format":"double"},"country":{"type":"string"},"currentStatus":{"type":"string","enum":["ACTIVE","INJURED","SUSPENDED","RETIRED"]},"registrationDate":{"type":"string","format":"date"},"victories":{"type":"integer","format":"int32"},"defeats":{"type":"integer","format":"int32"},"completedRaces":{"type":"integer","format":"int32"}}},"RaceRegistrationRequest":{"type":"object","properties":{"competitorId":{"type":"string","format":"uuid"},"teamId":{"type":"integer","format":"int64"},"startingPosition":{"type":"integer","format":"int32"},"registeredBy":{"type":"string","maxLength":120,"minLength":0}},"required":["registeredBy","startingPosition"]},"RaceRegistrationResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"raceId":{"type":"integer","format":"int64"},"raceName":{"type":"string"},"competitorId":{"type":"string","format":"uuid"},"competitorName":{"type":"string"},"teamId":{"type":"integer","format":"int64"},"teamName":{"type":"string"},"registrationDate":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["PENDING","APPROVED","REJECTED","CANCELLED"]},"startingPosition":{"type":"integer","format":"int32"},"validationNotes":{"type":"string"},"registeredBy":{"type":"string"}}},"RefreshRequest":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["password","username"]},"UpdateTeamStatusRequest":{"type":"object","properties":{"status":{"type":"string","enum":["ACTIVE","SUSPENDED","INACTIVE"]}},"required":["status"]},"RejectRegistrationRequest":{"type":"object","properties":{"reason":{"type":"string","maxLength":500,"minLength":0}},"required":["reason"]},"UpdateRaceStatusRequest":{"type":"object","properties":{"status":{"type":"string","enum":["DRAFT","OPEN_FOR_REGISTRATION","CLOSED_FOR_REGISTRATION","IN_PROGRESS","COMPLETED","CANCELLED"]}},"required":["status"]},"UpdateStatusRequest":{"type":"object","properties":{"status":{"type":"string","enum":["ACTIVE","INJURED","SUSPENDED","RETIRED"]}},"required":["status"]},"PageTeamResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/TeamResponse"}},"number":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"pageSize":{"type":"integer","format":"int32"},"pageNumber":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"CompetitorStandingResponse":{"type":"object","properties":{"rank":{"type":"integer","format":"int32"},"competitorId":{"type":"string","format":"uuid"},"name":{"type":"string"},"nickname":{"type":"string"},"points":{"type":"integer","format":"int32"},"victories":{"type":"integer","format":"int32"},"secondPlaces":{"type":"integer","format":"int32"},"thirdPlaces":{"type":"integer","format":"int32"},"racesWithResults":{"type":"integer","format":"int32"}}},"StandingsResponse":{"type":"object","properties":{"competitors":{"type":"array","items":{"$ref":"#/components/schemas/CompetitorStandingResponse"}},"teams":{"type":"array","items":{"$ref":"#/components/schemas/TeamStandingResponse"}}}},"TeamStandingResponse":{"type":"object","properties":{"rank":{"type":"integer","format":"int32"},"teamId":{"type":"integer","format":"int64"},"name":{"type":"string"},"points":{"type":"integer","format":"int32"},"victories":{"type":"integer","format":"int32"},"secondPlaces":{"type":"integer","format":"int32"},"thirdPlaces":{"type":"integer","format":"int32"},"racesWithResults":{"type":"integer","format":"int32"}}},"PageRaceResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/RaceResponse"}},"number":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PageCompetitorResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/CompetitorResponse"}},"number":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"UserProfileResponse":{"type":"object","properties":{"subject":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"fullName":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}}}},"securitySchemes":{"keycloak":{"type":"oauth2","description":"Usuarios de prueba del realm:\nadmin/Admin123! (todo),\norganizer/Organizer123! (carreras, inscripciones, resultados)\ny viewer/Viewer123! (solo lectura).","flows":{"authorizationCode":{"authorizationUrl":"http://localhost:8081/realms/camel-racing/protocol/openid-connect/auth","tokenUrl":"http://localhost:8081/realms/camel-racing/protocol/openid-connect/token","scopes":{}}}},"bearer-token":{"type":"http","description":"Pegá acá un access_token ya obtenido (sin la palabra\n\"Bearer\", Swagger la agrega sola). Por ejemplo, el que\ndevuelve POST /api/auth/login.","scheme":"bearer","bearerFormat":"JWT"}}}}